// src/app/model/ISesion.ts
interface ISesion {
    id_sesion?: number;
    id_usuario?: number;
    login?: string;
    nombre_apellido?: string;
    apellido_nombre?: string;
    id_perfil?: number;
    descripcion_perfil?: string;
    // nombre?: string;
    // apellido?: string;
    // email?: string;
    // estado_usuario?: string;
    // descripcio_estado_usuario?: string;
}

export default ISesion;